+++ vllm/worker/neuron_model_runner.py
@@ -5,7 +5,8 @@ import torch
 from vllm.config import (DeviceConfig, ModelConfig, ParallelConfig,
                          SchedulerConfig)
 from vllm.logger import init_logger
-from vllm.model_executor import get_model, SamplingMetadata
+from vllm.model_executor import SamplingMetadata
+from vllm.model_executor.neuron_model_loader import get_model
 from vllm.sampling_params import SamplingParams, SamplingType
 from vllm.sequence import SamplerOutput, SequenceData, SequenceGroupMetadata
 from vllm.utils import (async_tensor_h2d, pin_memory_available,
